change the type of mtu to u16 (#38)
* change the type of mtu to u16

+++ examples/read-async.rs
@@ -47,7 +47,7 @@ async fn main_entry(
     });
 
     let mut dev = tun2::create_as_async(&config)?;
-    let size = dev.as_ref().mtu()? + tun2::PACKET_INFORMATION_LENGTH;
+    let size = dev.as_ref().mtu()? as usize + tun2::PACKET_INFORMATION_LENGTH;
     let mut buf = vec![0; size];
     loop {
         tokio::select! {
